Re #1680: Fixed compile warnings on VS2005 for x64/win64 config.
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@4761 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjlib/src/pj/errno.c
@@ -109,7 +109,7 @@ static int pjlib_error(pj_status_t code, char *buf, pj_size_t size)
 
     len = pj_ansi_snprintf( buf, size, "Unknown pjlib error %d", code);
     if (len < 1 || len >= (int)size)
-	len = size - 1;
+	len = (int)(size - 1);
     return len;
 }
 
@@ -205,7 +205,7 @@ PJ_DEF(pj_str_t) pj_strerror( pj_status_t statcode,
     }
 
     if (len < 1 || len >= (int)bufsize) {
-	len = bufsize - 1;
+	len = (int)(bufsize - 1);
 	buf[len] = '\0';
     }
